Updating History
Polishing how to print the answer for rational fraction. Curing the fact that x(x-1) is understood a s the function x of (x-1) by math.js. Modifying the math.js files to remove the ~spaces in the implicit multiplication.
Adding the "once" option to avoid accumulating irrelevant event listener. (need to be checked fo all events).